+++ src/LASRapi/api.h
@@ -214,6 +214,7 @@ Pipeline reader_rectangles(std::vector<double> xmin, std::vector<double> ymin, s
 Pipeline region_growing(std::string connect_uid_raster, std::string connect_uid_seeds, double th_tree = 2, double th_seed = 0.45, double th_cr = 0.55, double max_cr = 20, std::string ofile = "");
 Pipeline remove_attribute(std::string name);
 Pipeline remove_attributes(std::vector<std::string> names);
+Pipeline remove_rgb();
 Pipeline keep_attributes(std::vector<std::string> names);
 Pipeline set_crs_epsg(int epsg);
 Pipeline set_crs_wkt(std::string wkt);